chore: Refractored ``runner_util`` function.
