+++ runner/runner.py
@@ -10,61 +10,12 @@ from plugins import IntegrityChecker, PipelineAnalyser
 from plugins.base import Plugin
 from type import Experiment
 from utils.flatten import flatten
-
+from utils.run_helpers import overwrite_model_configs_, append_parent_path_and_id_
 from .evaluation import evaluate
 from .store import Store
 
+
 obligatory_plugins = [PipelineAnalyser(), IntegrityChecker()]
-
-
-def overwrite_model_configs_(config: Experiment, pipeline: Pipeline) -> None:
-    """
-    Takes global config values and overwrites the config of
-    each model in the pipeline (if they have that attribute already).
-
-    Parameters
-    ----------
-    config
-        Configurations of the experiment
-    pipeline
-        The entire pipeline
-
-    Example
-    -------
-    if ``save_remote`` is set in the experiment configuration file,
-    it will overwrite ``save_remote`` in all models that have that attribute.
-
-    Returns
-    -------
-    None
-
-    """
-
-    for key, value in vars(config).items():
-        if value is not None:
-            for model in flatten(pipeline.children()):
-                if hasattr(model, "config"):
-                    if hasattr(model.config, key):
-                        vars(model.config)[key] = value
-
-
-def append_parent_path_and_id_(pipeline: Pipeline) -> None:
-    entire_pipeline = pipeline.dict_children()
-
-    def append(block, parent_path: str, id_with_prefix: str):
-        block["obj"].parent_path = parent_path
-        if not isinstance(block["obj"], DataSource):
-            block["obj"].id += id_with_prefix
-
-        if "children" in block:
-            for i, child in enumerate(block["children"]):
-                append(
-                    child,
-                    parent_path=f"{parent_path}/{block['obj'].id}",
-                    id_with_prefix=f"{id_with_prefix}-{i}",
-                )
-
-    append(entire_pipeline, Const.output_pipelines_path, "")
 
 
 class Runner:
